A significant refactor, introducing entry builder strategies.
Strategies increase the options available for building lorebook entries.  For now, this is just the box-standard stuff.

The whole `priorityDelta` and `searchDelta` stuff was extracted into the `DepthDelta` strategy.

A `Fixed` strategy is provided in case you just want to set a configuration once and inherit it across all nested entries.